Stability pool v2

+++ test/mocks/reward/accumulator/MockMultipleRewardCompoundingAccumulator.sol
@@ -2,12 +2,10 @@
 
 pragma solidity >=0.8.28 <0.9.0;
 
-// import {UUPSUpgradeable} from "@openzeppelin/contracts-upgradeable/proxy/utils/UUPSUpgradeable.sol";
 import {Initializable} from "@openzeppelin/contracts-upgradeable/proxy/utils/Initializable.sol";
 
 import {MultipleRewardCompoundingAccumulator} from "src/reward/accumulator/MultipleRewardCompoundingAccumulator.sol";
 
-// UUPSUpgradeable,
 contract MockMultipleRewardCompoundingAccumulator is Initializable, MultipleRewardCompoundingAccumulator {
     event AccumulateReward(address token, uint256 amount);
 
@@ -21,10 +19,7 @@ contract MockMultipleRewardCompoundingAccumulator is Initializable, MultipleRewa
     function initialize(address owner_) external initializer {
         _initializeOwner(owner_);
         __ReentrancyGuardTransient_init();
-        // __MultipleRewardCompoundingAccumulator_init();
     }
-
-    // function _authorizeUpgrade(address newImplementation) internal virtual override {}
 
     function setTotalPoolShare(uint256 _totalPoolShare, uint128 _product) external {
         totalPoolShare = _totalPoolShare;
@@ -38,7 +33,6 @@ contract MockMultipleRewardCompoundingAccumulator is Initializable, MultipleRewa
 
     function reentrantCall(bytes calldata _data) external nonReentrant {
         (bool _success, ) = address(this).call(_data);
-        // below lines will propagate inner error up
         if (!_success) {
             // solhint-disable-next-line no-inline-assembly
             assembly {
@@ -58,26 +52,17 @@ contract MockMultipleRewardCompoundingAccumulator is Initializable, MultipleRewa
         return (userProduct, userPoolShare);
     }
 
-    // expose some internal functions for testing
-    function tokenToExponentToIntegral(address token, uint8 exponent) public view returns (uint192 globalIntegral) {
-        globalIntegral = _tokenToExponentToIntegral(token, exponent);
+    function tokenToExponentToIntegral(address token, uint8 exponent) public view returns (uint256 globalIntegral) {
+        globalIntegral = uint256(_tokenToExponentToIntegral(token, exponent));
     }
 
-    /// @notice Get the user reward snapshot for a specific account and token.
-    /// @param account The address of user to query.
-    /// @param token The address of reward token to query.
-    /// @return timestamp The timestamp when the snapshot is updated
-    /// @return integral The reward integral until now.
-    /// @return pending The number of pending rewards.
-    /// @return claimed_ The number of claimed rewards.
-    /// @dev The integral is defined as 1e18 * ∫(rate(t) * prod(t) / totalPoolShare(t) dt).
     function userRewardSnapshot(
         address account,
         address token
-    ) public view returns (uint64 timestamp, uint192 integral, uint128 pending, uint128 claimed_) {
+    ) public view returns (uint64 timestamp, uint256 integral, uint128 pending, uint128 claimed_) {
         UserRewardSnapshot memory snapshot = _userRewardSnapshot(account, token);
         timestamp = snapshot.checkpoint.timestamp;
-        integral = snapshot.checkpoint.integral;
+        integral = uint256(snapshot.checkpoint.integral);
         pending = snapshot.rewards.pending;
         claimed_ = snapshot.rewards.claimed;
     }
